refactor(analysis): split content complexity cyclomatic tests
